refactor: 당첨 번호, 보너스 번호 유효성 검사 App 클래스에서 수행하도록 변경
