+++ index.html
@@ -29,7 +29,7 @@
                   required
                 />
               </div>
-              <button class="purchase-btn">구입</button>
+              <button type="submit" class="purchase-btn">구입</button>
             </form>
           </div>
           <div class="lotto-display"></div>
